tests: add pseudo selector testing page
